server tooling: wider ruff net, strict pytest config
ruff select grows SIM/RUF/ASYNC/PT (ASYNC and RUF006-family now police
the blocking-call and dangling-task classes this pass fixed by hand),
with per-file ignores replacing a dozen inline noqa in conftest and
excusing the test suite's deliberate sim-polling and composite-assert
style. Findings fixed: contextlib.suppress in the runner kill path,
collapsed elif chains in the sim, itertools.pairwise in tests, dead
counters and redundant int() casts. pytest gains --strict-markers (a
typo'd marker now fails instead of silently running), pytest-timeout
at 300 s, and the asyncio fixture-scope setting that silences the
pytest-asyncio deprecation warning.

+++ server/tests/test_building.py
@@ -138,7 +138,7 @@ def hover(cell, alt, drone_id="d0"):
 
 def run_place(tracker, drones, seconds, dt=0.1):
     all_placed, all_refused = [], []
-    for _ in range(int(round(seconds / dt))):
+    for _ in range(round(seconds / dt)):
         placed, refused = tracker.tick(drones, dt)
         all_placed.extend(placed)
         all_refused.extend(refused)
@@ -224,7 +224,7 @@ def test_place_refused_on_keep_out():
 
 def run_sources(drones, sources, carry, seconds, dt=0.1):
     pickups = []
-    for _ in range(int(round(seconds / dt))):
+    for _ in range(round(seconds / dt)):
         pickups.extend(tick_sources(drones, sources, carry, dt))
     return pickups
